+++ docs/CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 * Norwegian Bokmål translations. [Andreas Lorentsen]
 * Updated Afrikaans translations. [jcbrand]
 * Add responsiveness to CSS. We now use Sass preprocessor for generating CSS. [jcbrand]
+* #292 Better support for XEP-0085 Chat State Notifications. [jcbrand]
 
 0.8.6 (2014-12-07)
 ------------------